Update README.md
Works on Python. You need a file on your computer where you can enter your passwords, the number is not limited, a txt file. After starting the program, specify the path. The program was created for educational purposes and for checking your own passwords and addresses; I am not responsible for any illegal actions.